alpha for mesh edges
