+++ samples/gfx/edges.c
@@ -4,6 +4,7 @@
 #include "gfx/edges.h"
 
 #include "edge.glsl.h"
+#include "box3d/types.h"
 #include "gfx/geometry_registry.h"
 #include "gfx/scene_target.h"
 
@@ -36,9 +37,10 @@ EdgeOverlayParams GetDefaultEdgeParams( void )
 	p.showHeightfields = true;
 	p.thicknessPx = 1.5f;
 	p.zBias = 1.0e-6f;
-	p.convexColor = MakeVec4( 0.4f, 0.4f, 0.4f, 1.0f );
+	p.convexColor = MakeVec4( 1.0f, 1.0f, 1.0f, 0.75f );
 	p.concaveColor = MakeVec4( 0.50f, 0.10f, 0.10f, 1.0f );
-	p.flatColor = MakeVec4( 1.0f, 1.0f, 1.0f, 1.0f );
+	p.flatColor = MakeColor( b3_colorPaleGreen );
+	p.flatColor.w = 0.3f;
 	return p;
 }
