+++ samples/gfx/utility.h
@@ -18,6 +18,7 @@
 #pragma once
 
 #include "box3d/math_functions.h"
+#include "box3d/types.h"
 
 #include <math.h>
 
@@ -199,4 +200,19 @@ static inline Mat4 MakeMat4FromTransform( b3Transform t, b3Vec3 scale )
 	m.cz = MakeVec4( scale.z * r.cz.x, scale.z * r.cz.y, scale.z * r.cz.z, 0.0f );
 	m.cw = MakeVec4( t.p.x, t.p.y, t.p.z, 1.0f );
 	return m;
+}
+
+static inline float SrgbToLinear( float c )
+{
+	return c <= 0.04045f ? c / 12.92f : powf( ( c + 0.055f ) / 1.055f, 2.4f );
+}
+
+static inline Vec4 MakeColor( b3HexColor hexColor )
+{
+	uint32_t v = (uint32_t)hexColor;
+	float k = 1.0f / 255.0f;
+	float r = (float)( ( v >> 16 ) & 0xFFu ) * k;
+	float g = (float)( ( v >> 8 ) & 0xFFu ) * k;
+	float b = (float)( v & 0xFFu ) * k;
+	return MakeVec4( SrgbToLinear( r ), SrgbToLinear( g ), SrgbToLinear( b ), 1.0f );
 }
